% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Launch Domain Expertise Collector `Shiny` app to collect prior information}
\usage{
collect()
}
\value{
The `collect` function is used for the side effect of starting
the Domain Expertise Collector `Shiny` app
}
\description{
Run this function to start the Collector `Shiny` app.
}
\examples{
if (interactive()) {
  collect()
}
}
