% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{create_seq_from_q}
\alias{create_seq_from_q}
\title{Create dummy sequence of with `lt` and `ut` at `lq` and `uq`, for use in
`fitdistrplus` functions, which only take data, from which the qunatile
are concluded}
\usage{
create_seq_from_q(lq, uq, lt, ut)
}
\arguments{
\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{lt}{A single number between 0 and 1, representing the lower threshold}

\item{ut}{A single number between 0 and 1, representing the upper threshold}
}
\value{
A vector of 201 sequential numbers
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
create_seq_from_q(0.01, 0.99, 0.2, 0.8)
}
