% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{create_start}
\alias{create_start}
\title{Process starting value inputs in `Shiny`}
\usage{
create_start(dist, start = NULL)
}
\arguments{
\item{dist}{A single string, representing the distribution, either 'gamma'
or 'lognormal'}

\item{start}{A list with 4 elements named 'shape', 'rate', 'meanlog', and
'sdlog', containing starting values for the `fitdistrplus` functions for
Gamma and Log-Normal}
}
\value{
A smaller version of `start` that only contains parameters of the
distribution named in `dist`, or `NULL`, if all relevant parameters are 0
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
create_start(
  dist = "gamma",
  start = list(
    shape = 1,
    rate = 1,
    meanlog = 1,
    sdlog = 1
   )
 )
}
