% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par_01}
\alias{get_par_01}
\title{Derive parameters for variables in range 0 to 1}
\usage{
get_par_01(lq, uq, lt, ut, type01 = "1", meanbeta = 0.5)
}
\arguments{
\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{lt}{A single number between 0 and 1, representing the lower threshold}

\item{ut}{A single number between 0 and 1, representing the upper threshold}

\item{type01}{A single string, representing the shape for distributions in
range '0 to 1' with one of the thresholds equal to 0 or 1: '1' = 'Beta',
'2' = 'Half-Normal'}

\item{meanbeta}{A single number between 0 and 1, representing the mean of the
Beta distribution when for a parameter in range '0 to 1', both thresholds are
0 and 1}
}
\value{
A list with two elements, named 'dist' and 'param', with 'dist'
being a single string with the name of the distribution and 'param' being a
named list with one element per parameter
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
get_par_01(0.01, 0.99, 0.2, 0.4)
}
