% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par_0inf}
\alias{get_par_0inf}
\title{Derive parameters for variables in range 0 to Infinity}
\usage{
get_par_0inf(lq, uq, lt, ut, start = NULL, type0inf = "1")
}
\arguments{
\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{lt}{A single number between 0 and 1, representing the lower threshold}

\item{ut}{A single number between 0 and 1, representing the upper threshold}

\item{start}{A list with 4 elements named 'shape', 'rate', 'meanlog', and
'sdlog', containing starting values for the `fitdistrplus` functions for
Gamma and Log-Normal}

\item{type0inf}{A single string, representing the shape for distributions in
range '0 to Inf': '1' = 'Gamma', '2' = 'Log-Normal', '3' = 'Truncated Normal'}
}
\value{
A list with two elements, named 'dist' and 'param', with 'dist'
being a single string with the name of the distribution and 'param' being a
named list with one element per parameter
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
get_par_0inf(0.01, 0.99, 0.5, 3,
  start = list(
    shape = 1,
    rate = 1,
    meanlog = 1,
    sdlog = 1
 ))
}
