% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par_infinf}
\alias{get_par_infinf}
\title{Derive parameters for variables in range -Infinity to Infinity}
\usage{
get_par_infinf(lq, uq, lt, ut)
}
\arguments{
\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{lt}{A single number between 0 and 1, representing the lower threshold}

\item{ut}{A single number between 0 and 1, representing the upper threshold}
}
\value{
A list with two elements, named 'dist' and 'param', with 'dist'
being a single string with the name of the distribution and 'param' being a
named list with one element per parameter
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
get_par_infinf(0.01, 0.99, -2, 2)
}
