% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_beta_prior}
\alias{plot_beta_prior}
\title{Plot Beta distributions}
\usage{
plot_beta_prior(shape1, shape2, lq = 0.01, uq = 0.99)
}
\arguments{
\item{shape1}{A single postive number, representing parameter shape1 of the
Beta distribution}

\item{shape2}{A single postive number, representing parameter shape1 of the
Beta distribution}

\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}
}
\value{
A plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_beta_prior(2, 2, lq = 0.01, uq = 0.99)
}
