% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_empty_prior}
\alias{plot_empty_prior}
\title{Plot an empty canvas, used when inputs lead to errors}
\usage{
plot_empty_prior()
}
\value{
An empty plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_empty_prior()
}
