% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gamma_prior}
\alias{plot_gamma_prior}
\title{Plot Gamma distributions}
\usage{
plot_gamma_prior(shape, rate, lq = 0.01, uq = 0.99)
}
\arguments{
\item{shape}{A single postive number, representing parameter shape of the
Gamma distribution}

\item{rate}{A single postive number, representing parameter rate of the
Gamma distribution}

\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}
}
\value{
A plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_gamma_prior(7.3, 5)
}
