% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_halfnormal_prior}
\alias{plot_halfnormal_prior}
\title{Plot Half-Normal distributions}
\usage{
plot_halfnormal_prior(mean, sd, lq = 0.01, uq = 0.99, zero_one = FALSE)
}
\arguments{
\item{mean}{A single number, representing parameter mean of the
Normal distribution}

\item{sd}{A single postive number, representing parameter sd of the
Normal distribution}

\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{zero_one}{A single boolean value, indicating if the plot is used for
a distribution in rang '0 to 1' (TRUE) or not (FALSE; default)}
}
\value{
A plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_halfnormal_prior(0, 0.2)
}
