% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lognormal_prior}
\alias{plot_lognormal_prior}
\title{Plot Log-Normal distributions}
\usage{
plot_lognormal_prior(meanlog, sdlog, lq = 0.01, uq = 0.99)
}
\arguments{
\item{meanlog}{A single number, representing parameter meanlog of the
Log-Normal distribution}

\item{sdlog}{A single postive number, representing parameter sdlog of the
Log-Normal distribution}

\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}
}
\value{
A plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_lognormal_prior(0.5, 0.2)
}
