% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior}
\alias{plot_prior}
\title{Create the plot for the Domain Expertise Collector `Shiny` app}
\usage{
plot_prior(
  range,
  lq,
  uq,
  lt,
  ut,
  start = NULL,
  type01 = "1",
  type0inf = "1",
  meanbeta = 0.5
)
}
\arguments{
\item{range}{A single string, representing the range of possible values the
parameter can take: '1' = '0 to 1', '2' = '0 to Inf', '3' = '-Inf to Inf'}

\item{lq}{A single number between 0 and 1, representing the quartile which
should have the value of the lower threshold}

\item{uq}{A single number between 0 and 1, representing the quartile which
should have the value of the upper threshold}

\item{lt}{A single number between 0 and 1, representing the lower threshold}

\item{ut}{A single number between 0 and 1, representing the upper threshold}

\item{start}{A list with 4 elements named 'shape', 'rate', 'meanlog', and
'sdlog', containing starting values for the `fitdistrplus` functions for
Gamma and Log-Normal}

\item{type01}{A single string, representing the shape for distributions in
range '0 to 1' with one of the thresholds equal to 0 or 1: '1' = 'Beta',
'2' = 'Half-Normal'}

\item{type0inf}{A single string, representing the shape for distributions in
range '0 to Inf': '1' = 'Gamma', '2' = 'Log-Normal', '3' = 'Truncated Normal'}

\item{meanbeta}{A single number between 0 and 1, representing the mean of the
Beta distribution when for a parameter in range '0 to 1', both thresholds are
0 and 1}
}
\value{
A plot
}
\description{
This function is exported only for its use in the `Shiny` app. Use
`collect()` to start the app.
}
\examples{
plot_prior("1", 0.01, 0.99, 0.3, 0.7)
}
